New assignees and components should always be part of the change set,
even if the set introduces and then later removes the need for the
assignee/component. This is because the board changes are created
by the board on update, and we only know there that we are adding
a component/assignee for the first time. If we do not do this,
the client will run into trouble if a later change set references
this assignee/component since it will not have a copy of the
assignee/component.