Move issues in and out of the non-cached 'done' states.
Moving an issue into a 'done' state, deletes an issue from the board,
and appears as a delete in the change log.
Moving an issue into a 'non-done' state from a 'done' state, could be
seen as a create of an issue, but at this time was too tricky to do.
Instead it forces a full refresh of the board. I think this is fine
since moving into 'done' (which is at the end of the workflow) is the
normal case, moving back out of a 'done' state should be an exception to
the rule to fix mistakes etc.