Issue #14 Rework the board table so that the scrollbars are always displayed to  …
make it more user friendly for clients who do not have gesture
scrolling.

This has an issue in that when scrolling to the last states, in swimlane
view it 'overscrolls'. Meaning that when we get to the end of the table
we are still scrolling and it never ends. It appears to be related to
the swimlane headers.